issue #49 and update to config default wrapper
